% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_genome.R
\name{bin_genome}
\alias{bin_genome}
\alias{set_binGenome_path}
\title{Wrapper for binGenome module}
\usage{
set_binGenome_path(path)

bin_genome(
  input_dir,
  strand,
  annotation,
  output_dir,
  strand_pattern = "pos/neg",
  bins = NULL,
  grep_pattern = NULL,
  fixedBinSizeUpstream = NULL,
  fixedBinSizeDownstream = NULL,
  bedgraphNames = NULL,
  annotationNames = NULL,
  cores = NULL,
  normalize = FALSE,
  tmpDir = NULL
)
}
\arguments{
\item{path}{Path to the binGenome.sh executable.}

\item{input_dir}{A character string of path to the folder containing bedgraph files.}

\item{strand}{An integer in c(-1, 0, 1) to specify if and how the strandedness
of coverage files should be considered.
0: no filter | 1: regular stranded | -1: reverse stranded.
if \code{strand != 0}, it groups files by applying grep according to
the \code{strand_pattern}.}

\item{annotation}{A character string of path to annotation file in BED format.}

\item{output_dir}{Output folder for binGenome.sh}

\item{strand_pattern}{Identifiers for positive and negative strand to match
in the names of annotation files, separated by '/'.
First element corresponds to positive and second to negative. Default is "pos/neg".}

\item{bins}{Number of bins to bin the gene body.}

\item{grep_pattern}{Filter for a grep pattern in files. perl is TRUE.}

\item{fixedBinSizeUpstream, fixedBinSizeDownstream}{Specifies the parameters that will be
used for binning the downstream or upstream regions, respectively. Format: "binsize:binnumber".}

\item{bedgraphNames, annotationNames}{A character vector of names for bedgraph and annotation files.}

\item{cores}{Number of threads to use. Default: max(1, parallel::detectCores() - 2)}

\item{normalize}{Logical. If \code{TRUE}, an additional per-gene normalized version is written.}

\item{tmpDir}{Path to a tmp directory. Passed on to --tmpDir parameter of \code{genomeBinner.jar}
as well as set in \code{-Djava.io.tmpdir} if the provided executable is a jar file.}
}
\value{
Invisibly the path on successful execution.

Returns the \code{output_dir} invisibly.
}
\description{
Runs the binGenome.sh.
Prints the log of how many regions were too small for binning in each sample.
If stranded, pos and neg are merged and stripped from the name.
}
\details{
For binning, only files with .bedgraph or .bedgraph.gz extension in the \code{input_dir} are considered.
The function filters the files by "\\.bedgraph$", hence grep_pattern does not have to include the extension.

The \code{strand_pattern} is used literally, so user should not escape regex metacharacters.
For example, by strand_pattern = "+/-", plus is matched literally without any problem.

Jar file's output is saved to the \verb{out/} directory inside \code{output_dir}.

Since user quotas might be restricted on clusters, the jar file might often
crash due to low space. This is why it might help to set \code{-Djava.io.tmpdir}
to a directory where there is more allowance.
}
