% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparator_metrics.R
\name{compare_features}
\alias{compare_features}
\alias{sq_y_dist_bins}
\alias{mean_sqdist_commonFeatures}
\alias{wassersteinOverBins}
\title{Compare binmatrix objects}
\usage{
compare_features(binMatrix_before, binMatrix_after, returnFeatures = F)

sq_y_dist_bins(binMatrix_before, binMatrix_after, with_direction = FALSE)

mean_sqdist_commonFeatures(binMatrix_before, binMatrix_after)

wassersteinOverBins(binMatrix_before, binMatrix_after)
}
\arguments{
\item{binMatrix_before, binMatrix_after}{An object of class binMatrix}

\item{binMatrix_after}{An object of class binMatrix}

\item{with_direction}{A logical: Should the direction of the differences
also be considered? If TRUE the signs are assigned from the absolute differences:
\code{y_binMatrix_before - y_binMatrix_after}.}

\item{binMatrix_before}{An object of class binMatrix}
}
\value{
A 3-element named integer vector corresponding to the numbers of common
features (intersection), of features found only in first and second matrices,
respectively, with names \code{c("commonFeats", "onlyInBefore", "onlyInAfter")}.
If returnFeatures = TRUE, a list is returned with the features themselves instead
of the numbers.

\strong{sq_y_dist_bins} | A numeric vector of bin-wise differences in the y values
of the binmatrices as in their metagene plots.

\strong{mean_sqdist_commonFeatures} | A numeric vector of length 1.
Mean of the squared distances between the two binmatrices.

\strong{wassersteinOverBins} | A numeric vector of same length as bin number.
}
\description{
Compare binmatrix objects

Compare binmatrix objects
}
\keyword{internal}
