% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{create_binmatrix}
\alias{create_binmatrix}
\title{Create binMatrix objects}
\usage{
create_binmatrix(
  cov_folder,
  condition_greps,
  condition_mapper,
  aggregate_FUN,
  ids_to_subset = NULL,
  gsub_name = NULL,
  ignore = ".norm.coverage.csv",
  cores = 1
)
}
\arguments{
\item{cov_folder}{Coverage folder, where the binning results are}

\item{condition_greps}{Patterns for each condition (will match coverage file names)}

\item{condition_mapper}{See plot_metagene_experiment}

\item{aggregate_FUN}{See plot_metagene_experiment}

\item{annotation_grep}{A character string to filter the annotation/region.}
}
\value{
a list of aggregated binMatrix table results
}
\description{
Read from the coverage folder,
group by conditions and
aggregate results using the provided aggregation function.
}
