% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_genome.R
\name{decompress_bedgraph_gz}
\alias{decompress_bedgraph_gz}
\title{Decompress bedgraph.gz files}
\usage{
decompress_bedgraph_gz(input_dir, output_dir = NULL, grep_pattern = NULL)
}
\arguments{
\item{input_dir}{Path to the folder containing .bedgraph.gz files.}

\item{output_dir}{Path to the folder where the decompressed .bedgraph files will be saved.
If NULL (default), decompresses in input_dir.}

\item{grep_pattern}{Optional. Filter for a grep pattern in files. perl is TRUE.}
}
\value{
Invisibly, a vector of decompressed file paths.
}
\description{
Decompress bedgraph.gz files
}
