% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{getBinTable}
\alias{getBinTable}
\title{Collect binTables}
\usage{
getBinTable(
  folder,
  condition,
  gsub_name = NULL,
  ids_to_subset = NULL,
  ignore = NULL,
  cores = 1
)
}
\arguments{
\item{folder}{Path to the directory of coverage files.}

\item{condition}{A character string to filter the condition/sample.}

\item{gsub_name}{A character string to remove from the filename after parsing.}

\item{ids_to_subset}{A character vector to subset regions}

\item{ignore}{A character string to filter out files. Generally used for .norm files.}

\item{annotation}{A character string to filter the annotation/region.}
}
\value{
a list of binTables.
}
\description{
Collect binTables
}
