% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_windows.R
\name{ids_from_gtf}
\alias{ids_from_gtf}
\title{extract ids from a GTF/GFF using filter expressions}
\usage{
ids_from_gtf(input_file, filter = filter_gtf_by(), id_column = "gene_id")
}
\arguments{
\item{input_file}{path to a GTF or GFF}

\item{filter}{Filtering expressions generated by filter_gtf_by().
Used to subset GTF entries before window generation.
Defaults to \code{\link{filter_gtf_by()}} (no filtering).}

\item{id_column}{Default \code{NULL}, returns ids in the same format as \code{\link{make_windows}} generates. This output can be used for \code{ids_to_subset}.
Set a gtf column name (e.g. \code{"gene_name"}) to get values of that column instead.}
}
\value{
A character vector of unique ids matching the filtering criteria.
}
\description{
applies the same filter syntax as \code{\link{filter_gtf_by}()} and returns character vector of matching ids,
ready to use for \code{ids_to_subset} in \code{\link{plot_metagene_experiment}}
}
\examples{
gtf <- system.file("extdata", "data", "example.gtf", package = "metageneR")
ids <- ids_from_gtf(gtf, filter = filter_gtf_by(type == "gene"))
# plot_metagene_experiment(..., ids_to_subset = ids)
}
