% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{name_output_with}
\alias{name_output_with}
\title{Utility function: naming output with parameters}
\usage{
name_output_with(enclosing_dir, pars, ext)
}
\arguments{
\item{enclosing_dir}{A character string indicating the directory of the file.}

\item{pars}{A named list of character string arguments. The names correspond to the parameters,
the values to the arguments.}

\item{ext}{A character string of extension.}
}
\value{
A character vector of file path generated by the parameters.
}
\description{
Creates a standard name based on given parameters.
Originally created for use with make_windows, hence parameters like
input_file, upstream and downstream will be ignored.
}
\examples{
pars$make_windows$path_to_output <- name_output_with(out_dir, pars$make_windows, ".bed")
}
