% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_metagene_from_json}
\alias{plot_metagene_from_json}
\title{Re-plot using pars in json}
\usage{
plot_metagene_from_json(
  plot_dir,
  file_name,
  path_to_binMatrixList,
  json_file,
  ids_to_subset = NULL,
  ...
)
}
\arguments{
\item{plot_dir}{Where to save PDF of the plot}

\item{file_name}{File identifier. "metagene_" gets added as prefix.}

\item{json_file}{JSON file with a list of parameters to use in plotting}

\item{...}{Parameters to set/overwrite those in the JSON. Same as \code{plot_metagene_experiment()}.}

\item{path_to_binMatrix}{Path to the RDS object of aggregated binMatrices
that has been created by \code{plot_metagene_experiment()}.}
}
\description{
This function enables re-plotting of a metagene plot
by using parameters given in a JSON file and by specifying parameters
to overwrite those in the JSON.
The JSON file should have an attribute called binMatrixList_names, which
indicate the conditions to be compared.
}
