% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binGenome.lib.R
\name{plotGroup}
\alias{plotGroup}
\title{Create a shape plot for multiple binMatrix instances}
\usage{
plotGroup(
  binMatrixList,
  labels,
  title = "",
  name = "",
  smooth = NULL,
  showYLab = TRUE,
  palette = grDevices::rainbow(length(binMatrixList)),
  normByLibSize = TRUE,
  normByBinlength = TRUE,
  normLibsize = 10^9,
  normByShapeSum = FALSE,
  normByShapeMax = FALSE,
  aggregateFun = mean,
  verticalLinesOnTicks = NULL,
  fixedLabelsStart = NULL,
  fixedLabelsEnd = NULL,
  fixedLabelsStartTotalBins = 0,
  fixedLabelsEndTotalBins = 0,
  legendNCol = 1,
  ylimScale = 1,
  ylab = NULL,
  showXLab = TRUE,
  showLegend = TRUE,
  minYA = NULL,
  maxYA = NULL,
  legendPos = "top",
  performWilcoxTest = FALSE,
  wilcoxTestPVTransformFUN = defaultPvalueColorTransformer,
  cexAxis = 1.3,
  cexLab = 1.4,
  cexTitle = 1.6,
  shapeLineWidth = 1.75,
  ablineLineType = 3,
  cexLegend = 2,
  legendInset = 0.006,
  legendXIntersp = 0.8,
  legendYIntersp = 1,
  cexLegendSymbol = 3,
  legendSpacingBetween = 20,
  lineYlab = 3,
  lineTitle = 1,
  numberOfYTicks = NULL,
  asIsYlab = FALSE,
  shapeLineType = 1
)
}
\arguments{
\item{binMatrixList}{List of \code{binMatrix} instances for which the shapes should be plotted.}

\item{labels}{Body labels of the (scaled) X axis; assumes that labels are equally spaced
over all bins and that the first and last positions are labeled.}

\item{title}{(optional) Title of the shape panel. Default: \code{""}.}

\item{name}{(optional) Name added as prefix to the auto-generated Y-label. Default: \code{""}.}

\item{smooth}{(optional) Function to smooth the shapes (e.g. \code{smooth()}). Default: \code{NULL}.}

\item{showYLab}{(optional) Display Y label on the left side. Default: \code{TRUE}.}

\item{palette}{(optional) Colors used for the shapes in the same order as in \code{binMatrixList}.
Default: \code{grDevices::rainbow(length(binMatrixList))}.}

\item{normByLibSize}{(optional) Should the coverage be normalized by library size? Default: \code{TRUE}.}

\item{normByBinlength}{(optional) Should the coverage be normalized by bin length? Default: \code{TRUE}.}

\item{normLibsize}{(optional) Library size to normalize with. Default: 10^9.}

\item{normByShapeSum}{(optional) Should the coverage be normalized by the sum of the shape? Default: \code{FALSE}.}

\item{normByShapeMax}{(optional) Should the coverage be normalized by the maximum?}

\item{aggregateFun}{(optional) Function used to aggregate data of a bin. Default: \code{mean}.}

\item{verticalLinesOnTicks}{(optional) An integer vector. Draw vertical lines at these x-tick positions; default: \code{NULL}.}

\item{fixedLabelsStart}{(optional) X labels for fixed-length start bins; see \code{labels}. Default: \code{NULL}.}

\item{fixedLabelsEnd}{(optional) X labels for fixed-length end bins; see \code{labels}. Default: \code{NULL}.}

\item{fixedLabelsStartTotalBins}{(optional) Number of bins defined as fixed-length start bins. Default: \code{0}.}

\item{fixedLabelsEndTotalBins}{(optional) Number of bins defined as fixed-length end bins. Default: \code{0}.}

\item{legendNCol}{(optional) Number of columns in the legend. Default: \code{1}.}

\item{ylimScale}{(optional) Maximal y-value of all shapes is multiplied by this factor (allows zoom). Default: \code{1}.}

\item{ylab}{(optional) Custom y-axis label; if not set, it is determined by the applied normalization. Default: \code{NULL}.}

\item{showXLab}{(optional) Show or hide labels on x-axis tick marks. Default: \code{TRUE}.}

\item{showLegend}{(optional) Show or hide legend. Default: \code{TRUE}.}

\item{minYA}{(optional) Custom minimal y value (only with \code{maxYA}). Default: \code{NULL}.}

\item{maxYA}{(optional) Custom maximal y value (only with \code{minYA}). Default: \code{NULL}.}

\item{legendPos}{(optional) Position of the legend; parameter of \code{legend(...)} (e.g. \code{"top"}, \code{"topright"}). Default: \code{"top"}.}

\item{performWilcoxTest}{(optional) Perform position-wise paired Wilcoxon test and indicate corrected p-values with colors if two matrices are given. Default: \code{FALSE}.}

\item{wilcoxTestPVTransformFUN}{(optional) Function that transforms a list of p-values into a list of colors.
Default is \code{defaultPvalueColorTransformer}, which is a function that assigns colors based on specified p-value thresholds:
\preformatted{
  defaultPvalueColorTransformer(pvs, defaultColor = "lightgrey", pvThresholds = c(10^-3, 10^-10, 10^-15), colors = c("yellow", "orange", "red")).
  pvs: pvalues,
  defaultColor: (optional) default color if pvalue is not smaller or equal than any threshold,
  pvThresholds: (optional) thresholds that are required to assign a specific color (p <= t),
  colors: (optional) colors associated with the thresholds; must have the same length as pvThresholds
}}

\item{cexAxis}{(optional) Scaling factor for axis text. Default: \code{1.3}.}

\item{cexLab}{(optional) Scaling factor for axis labels. Default: \code{1.4}.}

\item{cexTitle}{(optional) Scaling factor for title. Default: \code{1.6}.}

\item{shapeLineWidth}{(optional) Line width for shapes. Default: \code{1.75}.}

\item{ablineLineType}{(optional) Line type for vertical lines. 0: blank, 1: solid, 2: dashed, 3: dotted, 4: dotdash, 5: longdash, 6: twodash. Default: \code{3}.}

\item{cexLegend}{(optional) Scaling factor for legend labels. Default: \code{2}.}

\item{legendInset}{(optional) Inset distance(s) from margins as a fraction of the plot region when legend is placed by keyword; parameter of \code{legend(...)}. Default: \code{0.006}.}

\item{legendXIntersp}{(optional) Character interspacing factor for horizontal spacing of legend. Default: \code{0.6}.}

\item{legendYIntersp}{(optional) Character interspacing factor for vertical spacing of legend. Default: \code{0.6}.}

\item{cexLegendSymbol}{(optional) Scaling factor for legend symbols. Default: \code{3}.}

\item{legendSpacingBetween}{(optional) Horizontal spacing between legend items. Default: \code{20}.}

\item{lineYlab}{(optional) Controls horizontal spacing of Y label. Default: \code{3}.}

\item{lineTitle}{(optional) Controls vertical spacing of title. Default: \code{1}.}

\item{numberOfYTicks}{(optional) Number of ticks on the Y-axis; automatically determined by default. Default: \code{NULL}.}

\item{asIsYlab}{(optional) If \code{TRUE}, Y label is taken as-is without modification. Default: \code{FALSE}.}

\item{shapeLineType}{(optional) Vector of line types for the shapes. 0: blank, 1: solid, 2: dashed, 3: dotted, 4: dotdash, 5: longdash, 6: twodash. Default: \code{1}.)}
}
\value{
A plot is generated; no return value.
}
\description{
This function creates a shape plot for a list of \code{binMatrix} instances,
displaying all shapes in a single plot window.
All the parameters after \code{title} can be set in the dot arguments of \code{plot_metagene_experiment()}.
}
\keyword{internal}
