% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_metagene_from_binmatrix}
\alias{plot_metagene_from_binmatrix}
\alias{args_for_metagene_plot}
\title{Metagene plot}
\usage{
plot_metagene_from_binmatrix(
  plot_dir,
  file_name,
  binMatrixList,
  pvThresholdsVec = c(0.05, 10^-3, 10^-5),
  ...
)

args_for_metagene_plot()
}
\arguments{
\item{plot_dir}{A character string. Where to save PDF output.}

\item{file_name}{A character string. File identifier.
Currently "metagene_" is automatically added as prefix.}

\item{binMatrixList}{List of binMatrix instances.}

\item{pvThresholdsVec}{A vector of numeric between 0 and 1 to use
as p-value thresholds in defaultPvalueColorTransformer for Wilcoxon test.}
}
\value{
Returns no value. Prints the parameters.
}
\description{
Basic plot drawing function wrapped from binGenome plotting library
}
\keyword{internal}
